boardを2つのクラスに分割
