+++ src/jp/ac/oit/igakilab/tasks/trello/model/TrelloBoard.java
@@ -2,18 +2,11 @@ package jp.ac.oit.igakilab.tasks.trello.model;
 
 import java.io.PrintStream;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
-public class TrelloBoard {
-	protected String id;
-	protected String name;
-	protected String desc;
-	protected String shortLink;
-	protected Set<String> memberIds;
-	protected boolean isClosed;
+import jp.ac.oit.igakilab.tasks.trello.TrelloBoardData;
 
+public class TrelloBoard extends TrelloBoardData{
 	protected List<TrelloCard> cards;
 	protected List<TrelloList> lists;
 
@@ -22,82 +15,17 @@ public class TrelloBoard {
 	}
 
 	public void init(){
-		id = null;
-		name = null;
-		desc = null;
-		shortLink = null;
-		memberIds = new HashSet<String>();
-		isClosed = false;
+		super.init();
 		cards = new ArrayList<TrelloCard>();
 		lists = new ArrayList<TrelloList>();
 	}
 
 	public void clear(){
-		id = null;
-		name = null;
-		desc = null;
-		shortLink = null;
-		memberIds.clear();
-		isClosed = false;
+		super.clear();
 		cards.clear();
 		lists.clear();
 	}
 
-	public String getId() {
-		return id;
-	}
-
-	public void setId(String id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getDesc() {
-		return desc;
-	}
-
-	public void setDesc(String desc) {
-		this.desc = desc;
-	}
-
-	public String getShortLink() {
-		return shortLink;
-	}
-
-	public void setShortLink(String shortLink) {
-		this.shortLink = shortLink;
-	}
-
-	public Set<String> getMemberIds() {
-		return memberIds;
-	}
-
-	public void clearMemberId(){
-		memberIds.clear();
-	}
-
-	public void addMemberId(String mid){
-		memberIds.add(mid);
-	}
-
-	public void removeMemberId(String mid){
-		memberIds.remove(mid);
-	}
-
-	public boolean isClosed() {
-		return isClosed;
-	}
-
-	public void setClosed(boolean isClosed) {
-		this.isClosed = isClosed;
-	}
 
 	public List<TrelloCard> getCards() {
 		return cards;
@@ -132,6 +60,7 @@ public class TrelloBoard {
 	public boolean containsCardById(String cid){
 		return getCardById(cid) != null;
 	}
+
 
 	public List<TrelloList> getLists() {
 		return lists;
@@ -197,6 +126,4 @@ public class TrelloBoard {
 			stream.println("\t" + card.toString());
 		}
 	}
-
-
 }
